set geth to serve websockets on 8545
This change is necessary because we're changing truffle's default provider to
be websockets. Truffle has a lot of tests that rely on the provider's port to be 8545. To keep
tests working we'd have to either 1) individually update the truffle-config for
each test using geth or 2) to use geth's default webprovider port. This change
opts for option 2.

update geth ports